Add getMinNumNotZero function and update GMeter
A new function, getMinNumNotZero, was added to Utils.ts to return the minimum non-zero number from a pair. This function was then applied in GMeter.ts for more accurate calculation of minimum count of usage per second, minute, hour, and day. The previous getMinNum function was replaced, which previously considered zeros in its calculations.